add: fix kotlin tests, fix ci
